+++ Documentation/Docxs/DevMaterials.md
@@ -1,0 +1,9 @@
+# Development materials
+
+## DevTeam Tools:
+- ### [Basic Movie Info](https://snap.stanford.edu/data/web-Movies.html)
+- ### [Dynamo dataset](https://snap.stanford.edu/data/web-Movies.html )
+
+## DevTeam Learning materials:
+[Yeah... theres a lot of stuff please wait for the moment]
+
