+++ Documentation/Docxs/WorkMethodology.md
@@ -1,0 +1,18 @@
+# Work Methodology
+
+We use Scrum methodology to keep track of the project, we are also using a local library of our learning materials in order to reduce the amount of time used in investigation. We will share this library once we finish the project [It´s too much]
+
+## Organizer tools
+- Notion: Ultimate Tasks
+- Github
+## Comunnication Tools
+- Whatsapp 
+- Discord
+- Slack
+- Teams
+
+## Tech Tools
+- Spring: Main framework
+- InteliJ: Most used IDE
+- Junit: Main testing tool
+- Github actions: Easier CI/CD using github
