+++ README.md
@@ -1,0 +1,26 @@
+# API REST Movie FANS Project
+
+![](Documentation/Images/PadawansCover.png)
+
+## Team Members :star:
+
+[Emilio Campos Daguer](CVs/CV_EmilioCampos.pdf)
+
+[Hector Abraham Valvieso Cosgalla]()
+
+[Josue Israel Canul Ordoñez](CVs/CV_JosueCanulpdf)
+
+[Luis Miguel Medina Avila](CVs/CV_LuisMedina.pdf)
+
+## Index :open_file_folder:
+
+* [Requirements](Documentation/Docxs/Requirements.md)
+* [Work Method & Tools](Documentation/Docxs/WorkMethodology.md)
+* [Diagrams](Documentation/Docxs/DiagramClasses.md)
+* [Links & Materials](Documentation/Docxs/DevMaterials.md)
+
+## Project Materials :closed_book:
+
+* [Material of presentation](/Files/First%20Stage.pdf)
+
+* [Video presentation](https://youtu.be/NaFzGoXHl6E)
