--- developer/src/test/java/com/develop/developer/DeveloperApplicationTests.java
@@ -1,13 +1,0 @@
-package com.develop.developer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DeveloperApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
