+++ movieproject/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Mon Nov 21 10:39:05 CST 2022
+gradle.version=7.5.1
