+++ movieproject/bin/main/application.properties
@@ -1,0 +1,1 @@
+
