+++ movieproject/src/main/java/com/develop/developer/DeveloperApplication.java
@@ -1,0 +1,13 @@
+package com.develop.developer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DeveloperApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DeveloperApplication.class, args);
+	}
+
+}
